# Conflicts:
#	src/Orchard.Web/Core/Navigation/Controllers/AdminController.cs

+++ src/Orchard.Web/Modules/Orchard.Indexing/Controllers/AdminController.cs
@@ -2,12 +2,12 @@
 using System.Linq;
 using System.Web.Mvc;
 using Orchard.Indexing.Services;
+using Orchard.Indexing.ViewModels;
 using Orchard.Localization;
 using Orchard.Logging;
 using Orchard.Security;
-using Orchard.Indexing.ViewModels;
 using Orchard.UI.Notify;
-using Orchard.Utility.Extensions;
+using static Orchard.Indexing.Helpers.IndexingHelpers;
 
 namespace Orchard.Indexing.Controllers {
     public class AdminController : Controller {
@@ -15,7 +15,7 @@ namespace Orchard.Indexing.Controllers {
         private readonly IIndexManager _indexManager;
 
         public AdminController(
-            IIndexingService indexingService, 
+            IIndexingService indexingService,
             IOrchardServices services,
             IIndexManager indexManager) {
             _indexingService = indexingService;
@@ -34,23 +34,21 @@ namespace Orchard.Indexing.Controllers {
                 IndexEntries = Enumerable.Empty<IndexEntry>(),
                 IndexProvider = _indexManager.GetSearchIndexProvider()
             };
-            
+
             if (_indexManager.HasIndexProvider()) {
                 viewModel.IndexEntries = _indexManager.GetSearchIndexProvider().List().Select(x => {
                     try {
                         return _indexingService.GetIndexEntry(x);
                     }
-                    catch(Exception e) {
+                    catch (Exception e) {
                         Logger.Error(e, "Index couldn't be read: " + x);
-                        return new IndexEntry { 
+                        return new IndexEntry {
                             IndexName = x,
                             IndexingStatus = IndexingStatus.Unavailable
                         };
                     }
                 });
             }
-            
-            // Services.Notifier.Information(T("The index might be corrupted. If you can't recover click on Rebuild."));
 
             return View(viewModel);
         }
@@ -59,7 +57,7 @@ namespace Orchard.Indexing.Controllers {
             if (!Services.Authorizer.Authorize(StandardPermissions.SiteOwner, T("Not allowed to manage the search index.")))
                 return new HttpUnauthorizedResult();
 
-            return View("Create", String.Empty);
+            return View("Create");
         }
 
         [HttpPost, ActionName("Create")]
@@ -68,7 +66,7 @@ namespace Orchard.Indexing.Controllers {
                 return new HttpUnauthorizedResult();
 
             var provider = _indexManager.GetSearchIndexProvider();
-            if (String.IsNullOrWhiteSpace(id) || id.ToSafeName() != id) {
+            if (!IsValidIndexName(id)) {
                 Services.Notifier.Error(T("Invalid index name."));
                 return View("Create", id);
             }
@@ -82,9 +80,9 @@ namespace Orchard.Indexing.Controllers {
                 provider.CreateIndex(id);
                 Services.Notifier.Success(T("Index named {0} created successfully", id));
             }
-            catch(Exception e) {
+            catch (Exception e) {
                 Services.Notifier.Error(T("An error occurred while creating the index: {0}", id));
-                Logger.Error("An error occurred while creatign the index " + id, e);
+                Logger.Error("An error occurred while creating the index " + id, e);
                 return View("Create", id);
             }
 
@@ -96,7 +94,12 @@ namespace Orchard.Indexing.Controllers {
             if (!Services.Authorizer.Authorize(StandardPermissions.SiteOwner, T("Not allowed to manage the search index.")))
                 return new HttpUnauthorizedResult();
 
-            _indexingService.UpdateIndex(id);
+            if (IsValidIndexName(id)) {
+                _indexingService.UpdateIndex(id);
+            }
+            else {
+                Services.Notifier.Error(T("Invalid index name."));
+            }
 
             return RedirectToAction("Index");
         }
@@ -106,7 +109,12 @@ namespace Orchard.Indexing.Controllers {
             if (!Services.Authorizer.Authorize(StandardPermissions.SiteOwner, T("Not allowed to manage the search index.")))
                 return new HttpUnauthorizedResult();
 
-            _indexingService.RebuildIndex(id);
+            if (IsValidIndexName(id)) {
+                _indexingService.RebuildIndex(id);
+            }
+            else {
+                Services.Notifier.Error(T("Invalid index name."));
+            }
 
             return RedirectToAction("Index");
         }
@@ -116,7 +124,12 @@ namespace Orchard.Indexing.Controllers {
             if (!Services.Authorizer.Authorize(StandardPermissions.SiteOwner, T("Not allowed to manage the search index.")))
                 return new HttpUnauthorizedResult();
 
-            _indexingService.DeleteIndex(id);
+            if (IsValidIndexName(id)) {
+                _indexingService.DeleteIndex(id);
+            }
+            else {
+                Services.Notifier.Error(T("Invalid index name."));
+            }
 
             return RedirectToAction("Index");
         }
